feat(tiktok): add Camoufox integration tests
Add a new integration test module exercising the Camoufox-based
TikTok download flow. The tests cover resolution, end-to-end
download and inspection of the downloaded file to ensure it
contains a video track, and verification that the Camoufox
strategy reports the expected name.

Replace the previous generic TikTok integration test with a
specialized suite that forces the Camoufox strategy via a patch
and uses the real TikTok demo URL. Introduce a helper to scan
downloaded files for a video track marker and assert download
metadata (content-type, size, and URL shape). Mark tests as
integration and slow, and require the scraping fixture.
